transcription-service/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
werkzeug>=2.3.0